refactor: rename `erlab.io.utilities` to `erlab.io.utils`
